refactor(domain && infra): changing bcrypt imports at test files
