+++ test/infrastructure/services/bcrypt_service_implementation.test.ts
@@ -1,5 +1,5 @@
 import { BcryptServiceImplementation } from "../../../src/infrastructure/services/bcrypt_service_implementation";
-import { BcryptService } from "../../../src/domain/contracts/adapters/bcrypt_service";
+import { BcryptService } from "../../../src/domain/contracts/services/bcrypt_service";
 
 describe('BcryptService', () => {
     let service: BcryptService;
